# Custom semaphore implementation