--- a/src/lib/shim/binary_spinning_sem.cc
+++ b/src/lib/shim/binary_spinning_sem.cc
@@ -1,5 +1,7 @@
 #include "binary_spinning_sem.h"
 
+#include "shadow_sem.h"
+
 #include <assert.h>
 #include <cstring>
 #include <errno.h>
@@ -7,23 +9,23 @@
 #include <unistd.h>
 
 BinarySpinningSem::BinarySpinningSem(ssize_t spin_max) : _thresh(spin_max) {
-    sem_init(&_semaphore, 1, 0);
+    shadow_sem_init(&_semaphore, 1, 0);
 }
 
 void BinarySpinningSem::post() {
-    sem_post(&_semaphore);
+    shadow_sem_post(&_semaphore);
     sched_yield();
 }
 
 void BinarySpinningSem::wait(bool spin) {
     if (spin) {
         for (int i = 0; _thresh < 0 || i < _thresh; ++i) {
-            if (sem_trywait(&_semaphore) == 0) {
+            if (shadow_sem_trywait(&_semaphore) == 0) {
                 return;
             }
         }
     }
-    sem_wait(&_semaphore);
+    shadow_sem_wait(&_semaphore);
 }
 
-int BinarySpinningSem::trywait() { return sem_trywait(&_semaphore); }
+int BinarySpinningSem::trywait() { return shadow_sem_trywait(&_semaphore); }