This adds a reimplementation of libc's semaphore interface. Using it instead of libc's semaphore in our binary_spinning_semaphore results in the underlying `syscall` calls going through the shim's overridden `syscall` function, which is whitelisted in the seccomp filter when seccomp is in use. i.e. this allows us to avoid unnecessary seccomp traps in the shim <-> shadow IPC (without the previous workaround of whitelisting *all* futex operations).
 
We also use our implementation of libc's semaphore to override libc's semaphore in the shim.

Progress on #1454