@sporksmith 2021-06-26
> @sporksmith do you want me to take a stab at the `shadow_sem.c` to `.cc` rewrite? Or is the baton back in your hands?

@RWails I'm comfortable doing it, but if you have cycles and would like to take a stab at it I won't turn down the help :)