@robgjansen 2021-06-29
Wouldn't the effort required to write in cpp be better spent writing in rust?