@RWails 2021-06-29
> Wouldn't the effort required to write in cpp be better spent writing in rust?

@sporksmith WDYT? The C -> CPP translation would be almost completely substitutional (I think), whereas a Rust implementation would be more substantial. Maybe still worth it if the module will eventually be rewritten in Rust, anyways.