@sporksmith 2021-06-29
I agree it doesn't make sense to block this on a Rust rewrite right now. I probably should've written it in Rust in the first place, but I initially thought it'd only be a few lines of C :). (It was, until I saw there was a performance degradation and had to fix it by avoiding unnecessary FUTEX_WAKE operations).

Actually, I think we've settled on dropping support for Centos 7, in which case I can just use the C11 atomics here, which should be a quite simple change. I'll send out the "drop centos 7 support" PR out and once it's finalized I'll plan to circle back to this and convert either to C11 or C++ atomics assuming it's not much work.

If it ends up being more substantial then I'd suggest we move on with the current use of gcc intrinsic atomics  for the moment, and circle back to rewrite it in Rust.