@sporksmith 2021-06-30
Some mildly ad-hoc measurements for posterity, using shadow config:
```
general:
  stop_time: 4 s
  bootstrap_end_time: 4 s
network:
  graph:
    type: 1_gbit_switch
hosts:
  host:
    quantity: 10
    processes:
    - path: /home/jnewsome/projects/phantom-paper/experiments/mezzo-phold/phold/target/release/phold
      args: 10 host uniform 0 100 0 1337
      start_time: 1 s
```

@ main: using libc semaphore; getting good performance by allowing all `futex` syscalls in our seccomp filter; however that exception isn't compatible with running threaded virtual processes, since we need to emualte their `futex` syscalls.
```
$ hyperfine -M3 'cd `mktemp -d` && shadow /home/jnewsome/projects/phantom-paper/experiments/mezzo-phold/config.yaml --interpose-method=preload -p 12 > shadow.log'
Benchmark #1: cd `mktemp -d` && shadow /home/jnewsome/projects/phantom-paper/experiments/mezzo-phold/config.yaml --interpose-method=preload -p 12 > shadow.log
  Time (mean ± σ):     21.646 s ±  0.352 s    [User: 103.112 s, System: 72.352 s]
  Range (min … max):   21.355 s … 22.038 s    3 runs
```

@ main, with the (thread-incompatible) futex exception in the seccomp filter disabled; significant performance penalty due to going through our SIGSYS signal handler every time the shadow IPC semaphore makes a futex syscall:
```
$ hyperfine -M3 'cd `mktemp -d` && shadow /home/jnewsome/projects/phantom-paper/experiments/mezzo-phold/config.yaml --interpose-method=preload -p 12 > shadow.log'
Benchmark #1: cd `mktemp -d` && shadow /home/jnewsome/projects/phantom-paper/experiments/mezzo-phold/config.yaml --interpose-method=preload -p 12 > shadow.log
  Time (mean ± σ):     25.668 s ±  0.507 s    [User: 115.138 s, System: 93.088 s]
  Range (min … max):   25.140 s … 26.151 s    3 runs
```

this PR @ 73b13f2da6d9c7a63f0f955e771be446e1f26b4a ("use c11 atomics"): Switches the shadow IPC to use the semaphore implementation in the PR. Because it uses an interposable `syscall` *library* call, our seccomp filter knows to let it pass. This appears to get back to the original performance using the glibc semaphore, but without having to let all futex syscalls through in our seccomp filter.
```
$ hyperfine -M3 'cd `mktemp -d` && shadow /home/jnewsome/projects/phantom-paper/experiments/mezzo-phold/config.yaml --interpose-method=preload -p 12 > shadow.log'
Benchmark #1: cd `mktemp -d` && shadow /home/jnewsome/projects/phantom-paper/experiments/mezzo-phold/config.yaml --interpose-method=preload -p 12 > shadow.log
  Time (mean ± σ):     21.648 s ±  0.450 s    [User: 103.053 s, System: 72.515 s]
  Range (min … max):   21.145 s … 22.012 s    3 runs
```

this PR @ 73b13f2da6d9c7a63f0f955e771be446e1f26b4a ("use c11 atomics"), but without the 'nwaiters' opt. The first version I implemented looked more like this, but there is a slight performance penalty vs the glibc semaphore; probably enough to justify the extra complexity. It's a little tricky though, and I wouldn't be shocked if there's some incorrect detail that can lead to deadlock. In a pinch we can disable it by removing the `nwaiters` check in `shadow_sem_post`.
```
$ hyperfine -M3 'cd `mktemp -d` && shadow /home/jnewsome/projects/phantom-paper/experiments/mezzo-phold/config.yaml --interpose-method=preload -p 12 > shadow.log'
Benchmark #1: cd `mktemp -d` && shadow /home/jnewsome/projects/phantom-paper/experiments/mezzo-phold/config.yaml --interpose-method=preload -p 12 > shadow.log
  Time (mean ± σ):     22.592 s ±  0.504 s    [User: 107.018 s, System: 76.233 s]
  Range (min … max):   22.198 s … 23.160 s    3 runs
```